+++ source/Analysis/GoodnessOfFit.vb
@@ -143,8 +143,8 @@ Friend Class GoodnessOfFit
         Dim ts_o_nan As TimeSeries = ts_o.Clone()
         datesToUpdate = New List(Of DateTime)
         For Each t As DateTime In ts_o_nan.Dates
-            For Each NaNPeriod As (start As DateTime, [end] As DateTime, count As Integer) In ts_s.NaNPeriods
-                If t >= NaNPeriod.start And t <= NaNPeriod.end Then
+            For Each NaNPeriod As (range As DateRange, count As Integer) In ts_s.NaNPeriods
+                If t >= NaNPeriod.range.start And t <= NaNPeriod.range.end Then
                     datesToUpdate.Add(t)
                     Exit For
                 End If
@@ -157,8 +157,8 @@ Friend Class GoodnessOfFit
         Dim ts_s_nan As TimeSeries = ts_s.Clone()
         datesToUpdate = New List(Of DateTime)
         For Each t As DateTime In ts_s_nan.Dates
-            For Each NaNPeriod As (start As DateTime, [end] As DateTime, count As Integer) In ts_o.NaNPeriods
-                If t >= NaNPeriod.start And t <= NaNPeriod.end Then
+            For Each NaNPeriod As (range As DateRange, count As Integer) In ts_o.NaNPeriods
+                If t >= NaNPeriod.range.start And t <= NaNPeriod.range.end Then
                     datesToUpdate.Add(t)
                     Exit For
                 End If