+++ source/CHANGELOG.md
@@ -13,6 +13,7 @@ FIXED:
 API-CHANGES:
 * New property `TimeSeries.NaNPeriods()` for retrieving a list of time periods consisting of NaN nodes
 * New method `TimeSeries.UpdateNode()` for updating the value of an existing node
+* New class `DateRange` representing a date range with defined start and end dates
 
 Version 2.7.1
 -------------